LayoutTests/platform/ios/http/tests/quicklook/top-navigation-blocked-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/quicklook/top-navigation-blocked.html' from frame with URL 'http://127.0.0.1:8000/quicklook/resources/top-navigation-blocked.docx'. The frame attempting navigation of the top-level window is sandboxed, but the 'allow-top-navigation' flag is not set.

CONSOLE MESSAGE: SecurityError (DOM Exception 18): The operation is insecure.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Run test
